raster/r.univar2/r.univar_main.c (process_raster): Added proper casts to `double' within the `CELL' and `FCELL' branches.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@30384 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.univar2/r.univar_main.c
@@ -211,7 +211,7 @@ process_raster (univar_stat *stats, int fd,
 		const int val_i = *((CELL *) ptr);
 
 		stats->sum += val_i;
-		stats->sumsq += (val_i * val_i);
+		stats->sumsq += (double) val_i * val_i;
 		stats->sum_abs += abs(val_i);
 
 		if (param.extended->answer)
@@ -233,7 +233,7 @@ process_raster (univar_stat *stats, int fd,
 		const float val_f = *((FCELL *) ptr);
 
 		stats->sum += val_f;
-		stats->sumsq += (val_f * val_f);
+		stats->sumsq += (double) val_f * val_f;
 		stats->sum_abs += fabs(val_f);
 
 		if (param.extended->answer)
